requirementsNew.txt
absl-py==1.4.0
aiofiles==0.7.0
asgiref==3.4.1
asttokens==2.0.5
astunparse==1.6.3
backcall 
backports.functools-lru-cache 
biwrap==0.1.6
build==0.10.0
cachetools==5.3.0
certifi 
charset-normalizer==2.0.12
click==8.0.1
colorama 
cycler==0.11.0
Cython==0.29.33
debugpy 
decorator 
executing==0.8.3
fastapi==0.68.1
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h11==0.12.0
h5py==3.1.0
httptools==0.2.0
idna==3.3
imageio==2.26.0
imgaug==0.4.0
importlib-metadata 
ipykernel 
ipython==8.4.0
jedi==0.18.1
jupyter_client 
jupyter_core 
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
lazy_loader==0.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.2
matplotlib-inline==0.1.3
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
nest-asyncio 
networkx==3.0
numpy==1.24.2
oauthlib==3.2.2
onnx==1.13.1
opencv-python==4.6.0.66
opencv-python-headless==4.5.3.56
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso 
pexpect==4.8.0
pickleshare 
Pillow==9.1.1
pip-tools==6.12.3
platformdirs 
prompt-toolkit==3.0.29
protobuf==3.20.3
psutil 
ptyprocess==0.7.0
pure-eval 
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
Pygments==2.12.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
python-dateutil 
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2022.1
PyWavelets==1.4.1
pywin32==227
PyYAML==5.4.1
pyzmq 
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scipy==1.8.1
seaborn==0.11.2
sentry-sdk==1.16.0
shapely==2.0.1
six==1.15.0
smmap==5.0.0
stack-data==0.3.0
starlette==0.14.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.5.0
tensorflow-plot==0.3.2
termcolor==1.1.0
tf-slim==1.1.0
thop==0.1.1.post2209072238
tifffile==2023.2.28
tomli==2.0.1
torch==1.8.2+cu102
torchaudio==0.8.2
torchvision==0.9.2+cu102
tornado 
tqdm==4.64.0
traitlets==5.3.0
typing-extensions==3.7.4.3
ultralytics==8.0.53
urllib3==1.26.15
uvicorn==0.15.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.0
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.12.1
zipp 
